Tracks collector (for local testing)
I wrote a simple collector to emulate the behavior of the Amazon EMR
collector that writes mapper output to temporary files, grouped by key,
and uses them to feed the reducer.  It was not necessary to use with the
process mapper/reducer, so it has been slightly hardcoded for the
pagerank step.