Finished C++ implementation
I implemented the process reducer in C++, and in order to do so, also
had to implement a min-max heap (so that I could use it to implement a
fixed-size heap), which seemed like a more efficient/scalable solution
that building a potentially massive vector & sorting it.  Local testing
did not reveal much of a difference, although it seemed that the sort
was a bit slower.  Since the heap has a fixed size, producing the top 20
ranks should be O(n * log(20)) instead of O(n * log n).